Refactor and bug fix for no nerd stat edge case
move getScorePlot inside plt.ts (i think this is better design)
